+++ src/base/string_attr_type.cc
@@ -33,7 +33,8 @@
 
 constexpr string_attr_type<void> SA_ORIGINAL_LINE("original_line");
 constexpr string_attr_type<void> SA_BODY("body");
-constexpr string_attr_type<void> SA_HIDDEN("hidden");
+constexpr string_attr_type<ui_icon_t> SA_HIDDEN("hidden");
+constexpr string_attr_type<void> SA_REPLACED("replaced");
 constexpr string_attr_type<intern_string_t> SA_FORMAT("format");
 constexpr string_attr_type<void> SA_REMOVED("removed");
 constexpr string_attr_type<void> SA_PREFORMATTED("preformatted");