+++ src/readline_highlighters.cc
@@ -189,8 +189,10 @@ readline_command_highlighter_int(attr_line_t& al,
     ws_index = line.find(' ', sub.lr_start);
     auto command = line.substr(sub.lr_start, ws_index);
     if (ws_index != std::string::npos) {
-        alb.overlay_attr(line_range(sub.lr_start + 1, ws_index),
-                         VC_ROLE.value(role_t::VCR_KEYWORD));
+        auto has_prefix = al.al_string[sub.lr_start] == ':';
+        alb.overlay_attr(
+            line_range(sub.lr_start + (has_prefix ? 1 : 0), ws_index),
+            VC_ROLE.value(role_t::VCR_KEYWORD));
 
         if (RE_PREFIXES.find_in(in_frag).ignore_error()) {
             lnav::snippets::regex_highlighter(