+++ src/styling.hh
@@ -86,11 +86,23 @@ struct icon_config {
 
 struct lnav_theme {
     std::map<std::string, scoped_value_t> lt_vars;
+
     positioned_property<icon_config> lt_icon_hidden;
     positioned_property<icon_config> lt_icon_ok;
     positioned_property<icon_config> lt_icon_info;
     positioned_property<icon_config> lt_icon_warning;
     positioned_property<icon_config> lt_icon_error;
+
+    positioned_property<icon_config> lt_icon_log_level_trace;
+    positioned_property<icon_config> lt_icon_log_level_debug;
+    positioned_property<icon_config> lt_icon_log_level_info;
+    positioned_property<icon_config> lt_icon_log_level_stats;
+    positioned_property<icon_config> lt_icon_log_level_notice;
+    positioned_property<icon_config> lt_icon_log_level_warning;
+    positioned_property<icon_config> lt_icon_log_level_error;
+    positioned_property<icon_config> lt_icon_log_level_critical;
+    positioned_property<icon_config> lt_icon_log_level_fatal;
+
     positioned_property<style_config> lt_style_identifier;
     positioned_property<style_config> lt_style_text;
     positioned_property<style_config> lt_style_alt_text;
@@ -109,6 +121,7 @@ struct lnav_theme {
     positioned_property<style_config> lt_style_adjusted_time;
     positioned_property<style_config> lt_style_skewed_time;
     positioned_property<style_config> lt_style_offset_time;
+    positioned_property<style_config> lt_style_time_column;
     positioned_property<style_config> lt_style_file_offset;
     positioned_property<style_config> lt_style_invalid_msg;
     positioned_property<style_config> lt_style_status_title;